refactor: move stage commands to `reth stage`
