+++ bin/reth/src/lib.rs
@@ -13,8 +13,6 @@ pub mod cli;
 pub mod config;
 pub mod db;
 pub mod dirs;
-pub mod drop_stage;
-pub mod dump_stage;
 pub mod merkle_debug;
 pub mod node;
 pub mod p2p;
