+++ bin/reth/src/stage/drop.rs
@@ -42,11 +42,11 @@ pub struct Command {
     /// - goerli
     /// - sepolia
     #[arg(
-        long,
-        value_name = "CHAIN_OR_PATH",
-        verbatim_doc_comment,
-        default_value = "mainnet",
-        value_parser = genesis_value_parser
+    long,
+    value_name = "CHAIN_OR_PATH",
+    verbatim_doc_comment,
+    default_value = "mainnet",
+    value_parser = genesis_value_parser
     )]
     chain: Arc<ChainSpec>,
