+++ bin/reth/src/stage/dump/hashing_account.rs
@@ -1,20 +1,26 @@
-use crate::{dump_stage::setup, utils::DbTool};
+use super::setup;
+use crate::utils::DbTool;
 use eyre::Result;
 use reth_db::{database::Database, table::TableImporter, tables};
-use reth_primitives::StageCheckpoint;
+use reth_primitives::{BlockNumber, StageCheckpoint};
 use reth_provider::Transaction;
-use reth_stages::{stages::StorageHashingStage, Stage, StageId, UnwindInput};
+use reth_stages::{stages::AccountHashingStage, Stage, StageId, UnwindInput};
 use std::{ops::DerefMut, path::PathBuf};
 use tracing::info;
 
-pub(crate) async fn dump_hashing_storage_stage<DB: Database>(
+pub(crate) async fn dump_hashing_account_stage<DB: Database>(
     db_tool: &mut DbTool<'_, DB>,
-    from: u64,
-    to: u64,
+    from: BlockNumber,
+    to: BlockNumber,
     output_db: &PathBuf,
     should_run: bool,
 ) -> Result<()> {
     let (output_db, tip_block_number) = setup::<DB>(from, to, output_db, db_tool)?;
+
+    // Import relevant AccountChangeSets
+    output_db.update(|tx| {
+        tx.import_table_with_range::<tables::AccountChangeSet, _>(&db_tool.db.tx()?, Some(from), to)
+    })??;
 
     unwind_and_copy::<DB>(db_tool, from, tip_block_number, &output_db).await?;
 
@@ -33,7 +39,7 @@ async fn unwind_and_copy<DB: Database>(
     output_db: &reth_db::mdbx::Env<reth_db::mdbx::WriteMap>,
 ) -> eyre::Result<()> {
     let mut unwind_tx = Transaction::new(db_tool.db)?;
-    let mut exec_stage = StorageHashingStage::default();
+    let mut exec_stage = AccountHashingStage::default();
 
     exec_stage
         .unwind(
@@ -47,9 +53,7 @@ async fn unwind_and_copy<DB: Database>(
         .await?;
     let unwind_inner_tx = unwind_tx.deref_mut();
 
-    // TODO optimize we can actually just get the entries we need for both these tables
-    output_db.update(|tx| tx.import_dupsort::<tables::PlainStorageState, _>(unwind_inner_tx))??;
-    output_db.update(|tx| tx.import_dupsort::<tables::StorageChangeSet, _>(unwind_inner_tx))??;
+    output_db.update(|tx| tx.import_table::<tables::PlainAccountState, _>(unwind_inner_tx))??;
 
     unwind_tx.drop()?;
 
@@ -65,7 +69,7 @@ async fn dry_run(
     info!(target: "reth::cli", "Executing stage.");
 
     let mut tx = Transaction::new(&output_db)?;
-    let mut exec_stage = StorageHashingStage {
+    let mut exec_stage = AccountHashingStage {
         clean_threshold: 1, // Forces hashing from scratch
         ..Default::default()
     };
